Componentized javascript code
